requirements-agent.txt
# AI Agent Requirements
# Core ML and LLM
openai>=1.0.0
tiktoken>=0.5.0

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Text Processing
rapidfuzz>=3.0.0

# Web Scraping (existing)
requests>=2.31.0
beautifulsoup4>=4.12.0
urllib3>=2.0.0

# Utility
python-dotenv>=1.0.0
pydantic>=2.0.0 